switched money datatype to decimal after group discussion. will have to add logic to convert to stripe compatible format later. fixes schema and removes mistake migration which clobbered the product id in the order_items.

+++ db/migrate/20150312154812_create_orders.rb
@@ -1,10 +1,10 @@
 class CreateOrders < ActiveRecord::Migration
   def change
     create_table :orders do |t|
-      t.integer :taxes
-      t.integer :delivery_cost
-      t.integer :subtotal
-      t.integer :total
+      t.decimal :taxes
+      t.decimal :delivery_cost
+      t.decimal :subtotal
+      t.decimal :total
 
       t.timestamps null: false
     end
