+++ db/migrate/20150312170552_add_total_price_to_order_item.rb
@@ -1,5 +1,5 @@
 class AddTotalPriceToOrderItem < ActiveRecord::Migration
   def change
-	  add_column :order_items, :total_price, :integer
+	  add_column :order_items, :total_price, :decimal
   end
 end
